% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adaptive_imgt_mapping}
\alias{adaptive_imgt_mapping}
\title{Adaptive to IMGT mapping table}
\format{
A data frame with 3 variables: \code{adaptive}, \code{imgt},
\code{species}
\describe{
\item{adaptive}{adpative format}
\item{imgt}{imgt format}
\item{species}{speices}
}
}
\usage{
adaptive_imgt_mapping
}
\description{
Source: https://en.wikipedia.org/wiki/Arcade_game#List_of_highest-grossing_games
}
\keyword{datasets}
