% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolsets.R
\name{add.background.point}
\alias{add.background.point}
\title{for facet plot, add background point}
\usage{
add.background.point(all_tsne, cluster = "cluster", group = "group",
  cor1 = "X", cor2 = "Y", sample.order, group.order)
}
\arguments{
\item{all_tsne}{all_tsne}

\item{cluster}{cluster}

\item{group}{group}

\item{cor1}{x}

\item{cor2}{y}

\item{sample.order}{sample.order}

\item{group.order}{group.order}
}
\value{
a merged dataframe
}
\description{
for facet plot, add background point
}
\examples{
merged_df <- add.background.point(all_tsne, sample.order = sample.order, group.order = group.order)

}
