% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_matrix.R
\name{add_gender_info_to_annoDf}
\alias{add_gender_info_to_annoDf}
\title{add_gender_info_to_annoDf}
\usage{
add_gender_info_to_annoDf(exprM, annoDf)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}
}
\value{
a ggplot object
}
\description{
add_gender_info_to_annoDf
}
