% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_addNewSecurity.R
\name{addNewSecurity}
\alias{addNewSecurity}
\title{to be deprecated later on
 add security to existing rds file
 @export}
\usage{
addNewSecurity(ticker)
}
\description{
to be deprecated later on
 add security to existing rds file
 @export
}

