% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_x100.R
\name{as_x100}
\alias{as_x100}
\title{Percentage}
\usage{
as_x100(x, t, d = 2)
}
\arguments{
\item{x}{value}

\item{t}{total}

\item{d}{digits to round}
}
\value{
x as a percentage of t
}
\description{
Percentage
}
\examples{
x100(25, 200)

}
