% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{bas_color_palette}
\alias{bas_color_palette}
\title{Color Palette}
\format{
An object of class \code{list} of length 3.
}
\usage{
bas_color_palette
}
\description{
Color Palette
}
\keyword{datasets}
