% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bdhLastestToMatrix.R
\name{bdhLastestToMatrix}
\alias{bdhLastestToMatrix}
\title{return a list of matrix which comprise of a consolidated value matrix and one date matrix given list of bdh data}
\usage{
bdhLastestToMatrix(bdhList, timeSeries, validDays = Inf, cl)
}
\arguments{
\item{bdhList}{list of bdh data to be converted to matrix}

\item{timeSeries}{date or character vector that determine the resulting matrix's rownames}

\item{validDays}{number of days a bdh data is valid since it is published}

\item{cl}{cluster that will be used for computation}
}
\description{
return a list of matrix which comprise of a consolidated value matrix and one date matrix given list of bdh data
}

