% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bdhToMatrix.c.R
\name{bdhToMatrix.c}
\alias{bdhToMatrix.c}
\title{c implementation of the bdhToMatrix 
return a list of matrix which comprise of a consolidated value matrix and one date matrix given list of bdh data}
\usage{
bdhToMatrix.c(bdhList, timeSeries, validDays = 730, lag = 0,
  announcementDate = NULL)
}
\arguments{
\item{bdhList}{list of bdh data to be converted to matrix}

\item{timeSeries}{date or character vector that determine the resulting matrix's rownames}

\item{validDays}{number of CALENDAR days a bdh data is valid since it is published, for price data that only valid for that day, use 1 with lag = 0}

\item{lag}{number of TRADING day lag, default is 0; 1 mean lag by one trading day}

\item{cl}{cluster that will be used for computation}
}
\description{
c implementation of the bdhToMatrix 
return a list of matrix which comprise of a consolidated value matrix and one date matrix given list of bdh data
}
