% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_by_seurat}
\alias{clustering_by_seurat}
\title{clustering by seurat}
\usage{
clustering_by_seurat(rawCountM, clusterNum = 5, colors.use = NULL,
  genes = NULL, nCol = 4)
}
\arguments{
\item{rawCountM}{raw expression matrix from 10x (can't apply to smart-seq data)}

\item{clusterNum}{cluster number}

\item{colors.use}{colors used for clustering (should be more than the cluster num)}

\item{genes}{genes wanted to used in staining plot (FeaturePlot)}

\item{nCol}{column set in staining plot (FeaturePlot)}
}
\value{
a seurat object and some plots
}
\description{
clustering by seurat
}
\examples{
load("/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/Aug_30/analysis_4/gbm.4.pos.neg.ctrl.vcl.Rdata")
exprM_ctrl <- exprs(gbm_ctrl)
mm10 <- read.table("/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/Reeson_mar/mm10.feature.info", row.names = 1, header = F)
rownames(exprM_ctrl) <- mm10[rownames(exprM_ctrl),]$V2
rawCountM <- exprM_ctrl; mycolors <- brewer.pal(8,"Set2")
ENSgene <- c("Top2a","Cdk1","Aurkb","Myc","Snai2","Sox9","Sox10","Foxd3","Phox2b","Isl1","Plp1","Tubb3","Tubb4a","Elavl4","Ret","Fabp7")
new_seuset <- clustering_by_seurat(rawCountM, clusterNum=5, colors.use=mycolors, genes=ENSgene)
}
