% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_computeNAV.R
\name{computeNAV}
\alias{computeNAV}
\title{given a matrix of portfolio ? and price matrix, compute the NAV of a portoflio
to be rebuilt / deprecated later}
\usage{
computeNAV(portfolio, price, vwap)
}
\description{
given a matrix of portfolio ? and price matrix, compute the NAV of a portoflio
to be rebuilt / deprecated later
}
