% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_conformDF.R
\name{conformDF}
\alias{conformDF}
\title{convert one data.frame to the format of target data.frame
S stands for Same, U stands for Union, I stands for Intersect}
\usage{
conformDF(from, to, locf = F, type = "S")
}
\description{
convert one data.frame to the format of target data.frame
S stands for Same, U stands for Union, I stands for Intersect
}

