% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_createPortfolio.R
\name{createPortfolio}
\alias{createPortfolio}
\title{to be depercated later
given params : capital ... etc
and list of stock / vwap / turnover
output the resulting portofolio in a matrix}
\usage{
createPortfolio(params, stockList, vwap, turnover)
}
\description{
to be depercated later
given params : capital ... etc
and list of stock / vwap / turnover
output the resulting portofolio in a matrix
}

