% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dateRemap.R
\name{dateRemap}
\alias{dateRemap}
\title{to be deprecated later
given a data.frame of Fiscal date map it to actual date
was used to map eps , but doesnt seem very useful now}
\usage{
dateRemap(Fiscal_df, Actual_df)
}
\description{
to be deprecated later
given a data.frame of Fiscal date map it to actual date
was used to map eps , but doesnt seem very useful now
}
