% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{DEG_pseudotime}
\alias{DEG_pseudotime}
\title{pseudotime analysis based on clustering.}
\usage{
DEG_pseudotime(exprM, annoDf, sampleNum = 200, knn = 10,
  ctrl = "negt", comp = "post", topNum = 200)
}
\arguments{
\item{exprM}{normalized expression matrix contain all cells}

\item{annoDf}{the annotation of the cells (columns) of the exprM (must have a column "pseudotime")}

\item{sampleNum}{how many cells for sampling}

\item{knn}{how many cells for smoothing}
}
\value{
a dataframe to show the gene, log2(sumA/sumB)
}
\description{
pseudotime analysis based on clustering.
}
\examples{
save(exprM, annoDf, file="/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/result1/exprM_3_anno.Rdata")
annoDf$cellGroup <- substr(rownames(annoDf), 1, 4)
}
