% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dfToMatrix.R
\name{dfToMatrix}
\alias{dfToMatrix}
\title{given a data in long form , covert it to wide form
will improve the function later before being used again}
\usage{
dfToMatrix(df, fieldname, locf = F)
}
\description{
given a data in long form , covert it to wide form
will improve the function later before being used again
}

