% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_evalNAV.R
\name{evalNAV}
\alias{evalNAV}
\title{given a time seeries of NAV  and benchmark (optional),
compute the statistics of the time series}
\usage{
evalNAV(NAV, benchmark = NULL)
}
\description{
given a time seeries of NAV  and benchmark (optional),
compute the statistics of the time series
}
