% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolsets.R
\name{example.function}
\alias{example.function}
\title{description of the example function}
\usage{
example.function(param1)
}
\arguments{
\item{param1}{param1}
}
\value{
the return object
}
\description{
description of the example function
}
\examples{
example code

}
