% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_number.R
\name{find_number}
\alias{find_number}
\title{Find numbers in a string}
\usage{
find_number(string, digits = 2, all = FALSE)
}
\arguments{
\item{string}{String in which to look}

\item{digits}{Number of digits to match}

\item{all}{Logical. Wether or not to include all instances. If F, will return only the first one.}
}
\value{
Vector with the found numbers.
}
\description{
Finds numbers qithin a string. Can find any specified number of digits and can return either the first instances or all instances.
}
\examples{
find_number("There are 123 apples and 456 oranges", digits = 3, all = T)
find_number("My number is 42", all = TRUE)
find_number("No numbers here", digits = 1)
}
