% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShp.R
\name{findShp}
\alias{findShp}
\title{Find shapefile}
\usage{
findShp(path)
}
\arguments{
\item{path}{Where to find the shapefiles. 'D:/example/location'}
}
\value{
Vector of shapefiles (only .shp).
}
\description{
Get a list (vector) of .shp files in a directory.
}
\examples{
# findShp('D:/example/location')
}
