% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rid.R
\name{get_rid}
\alias{get_rid}
\title{Get rid of ...}
\usage{
get_rid(df, var.name)
}
\arguments{
\item{df}{data.frame}

\item{var.name}{name (or names with c()) to be dropped}
}
\value{
data.frame
}
\description{
Get rid of ...
}
\examples{
'i really dont wanna write rn'
}
