% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_getFutChain.R
\name{getFutChain}
\alias{getFutChain}
\title{given future code , list of months code , starting year
return the list of future code}
\usage{
getFutChain(futureCode, months, years)
}
\description{
given future code , list of months code , starting year
return the list of future code
}

