% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_getIndexMember.R
\name{getIndexMember}
\alias{getIndexMember}
\title{give the index ticker, and date (optional) return the list of stock wgt from bloomberg}
\usage{
getIndexMember(index, date)
}
\description{
give the index ticker, and date (optional) return the list of stock wgt from bloomberg
}

