% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_getMedian.R
\name{getMedian}
\alias{getMedian}
\title{given a data matrix, return the median time series}
\usage{
getMedian(data, minCount = 1)
}
\arguments{
\item{data}{data matrix containing stock data time series}

\item{minCount}{minimum effective count to calculate median}
}
\description{
given a data matrix, return the median time series
}

