% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibberish.R
\name{gibberish}
\alias{gibberish}
\title{Gibberish generator}
\usage{
gibberish(
  n = 15,
  low_alpha = T,
  upp_alpha = T,
  num = T,
  sym = T,
  blacklist = c()
)
}
\arguments{
\item{n}{length}

\item{low_alpha}{include lower alpha}

\item{upp_alpha}{include upper alpha}

\item{num}{include numeric}

\item{sym}{include ascii symbols}

\item{blacklist}{symbols to remove from character pool}
}
\value{
gibberish
}
\description{
Gibberish generator
}
\examples{
for (i in 1:50) gibberish(i)
}
