% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_to_commit.R
\name{git_to_commit}
\alias{git_to_commit}
\title{Print Local git Repositories to Synchronize}
\usage{
git_to_commit(path = "~")
}
\arguments{
\item{path}{the root directory from which all folders are recursively
inspected.}
}
\description{
This function checks if all git repositories in the \code{path} folder are
update-to-date. It prints the list of git repositories for which changes
need to be committed and pushed to remote hub.
}
\examples{
\dontrun{
git_to_commit(path = "~/GitHub")
}
}
