% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus_equal.R
\name{\%-=\%}
\alias{\%-=\%}
\title{`roperators`'s minus-equal}
\usage{
x \%-=\% y
}
\value{
Same logic as ropreators's `%-=%`
}
\description{
Literally the same as roperators's `%-=%`. Just did not want to call `library(roperators, include.only = '%-=%')` every time I wanted to use it, and since I do load `rtoolbox` every time it was simpler to have it here.
}
\examples{
 n <- 0
 for (i in 1:10) {
  n \%-=\% i
 }
}
