% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Magrittr's assignment pipe}
\usage{
lhs \%<>\% rhs
}
\value{
Same logic as magrittr's assignment pipe
}
\description{
Literally the same as magrittr's assignment pipe `%<>%`. Just did not want to call `library(magrittr, include.only = '%<>%')` every time I wanted to use it, and since I do load `rtoolbox` every time it was simpler to have it here.
}
\examples{
# Will do this later
}
