% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in there}
\usage{
... \%notin\% NA
}
\value{
T or F
}
\description{
Literally the opposite of %in% and exactly the same as using !, but it's more explicit and easier to read.
}
\examples{
# 'x' \%notin\% c('y', 'z')
}
