% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpf.R
\name{gsea.GO.genes.keyWords}
\alias{gsea.GO.genes.keyWords}
\title{Get the genes of interested GO terms by some key words}
\usage{
gsea.GO.genes.keyWords(organism = "hs", keyWords = "axon",
  returnAll = F)
}
\arguments{
\item{organism}{organism (hs and mm)}

\item{keyWords}{keyWords (for grey function, e.g. cell cycle|DNA replication)}

\item{returnAll}{if return all the GO terms}
}
\value{
a list with the GO terms and genes
}
\description{
Get the genes of interested GO terms by some key words
}
\examples{
tmpList <- toC_get_genes_of_GO_by_keyWords(organism = "hs", keyWords = "axon", returnAll = F)
tmpList <- toC_get_genes_of_GO_by_keyWords(organism = "hs", 
                                           keyWords = "cell cycle|DNA replication|cell division|segregation")

}
