% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolsets.R
\name{gsea.ID2gene}
\alias{gsea.ID2gene}
\title{transfer the ID string to gene vector or string}
\usage{
gsea.ID2gene(ID, organism = "hs", returnVector = T)
}
\arguments{
\item{ID}{the ID string, like 4171/4175/5422/4172}

\item{organism}{organism (hs and mm)}

\item{returnVector}{return vector or string}
}
\value{
a transfromed gene name list or string
}
\description{
transfer the ID string to gene vector or string
}
\examples{
tmpgenes <- gsea.ID2gene("4171/4175/5422/4172", organism="hs")

}
