% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{hello_world}
\alias{hello_world}
\title{Print "Hello world"}
\usage{
hello_world(to_print = "Hello world", excited = FALSE)
}
\arguments{
\item{to_print}{A character string giving the text the function will print}

\item{excited}{Logical value specifying whether to include an exclamation
point after the text}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function that, by default, prints "Hello world". You can
customize the text to print (using the \code{to_print} argument) and add
an exclamation point (\code{excited = TRUE}).
}
\examples{
hello_world()
hello_world(excited = TRUE)
hello_world(to_print = "Hi world")

}
