% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lag.matrix.R
\name{lag.matrix}
\alias{lag.matrix}
\title{lag a matrix by n days, fill the missing data with the closest one}
\usage{
\method{lag}{matrix}(mat, n = 1)
}
\description{
lag a matrix by n days, fill the missing data with the closest one
}

