% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_lag.matrix.R
\name{lagMatrix}
\alias{lagMatrix}
\title{lag a matrix by n days, fill the missing data with the closest one}
\usage{
lagMatrix(mat, n = 1)
}
\description{
lag a matrix by n days, fill the missing data with the closest one
}

