% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{marker_identification_by_correlation}
\alias{marker_identification_by_correlation}
\title{marker identification by correlation}
\usage{
marker_identification_by_correlation(exprM, topNum = 10,
  plotHeatmap = T, group.order = NULL)
}
\arguments{
\item{topNum}{top X markers}

\item{plotHeatmap}{whether to plot heatmap or not (default yes)}

\item{group.order}{specify the cluster order (after pseudotime)}

\item{seuset}{a seurat object after clustering}
}
\value{
a marker dataframe and the heatmap
}
\description{
marker identification by correlation
}
\examples{
exprM <- seuset_2@scale.data; anno <- seuset_2@ident
}
