% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{marker_identification_by_seurat}
\alias{marker_identification_by_seurat}
\title{marker identification by seurat}
\usage{
marker_identification_by_seurat(seuset, topNum = 10, plotHeatmap = T,
  group.order = NULL)
}
\arguments{
\item{seuset}{a seurat object after clustering}

\item{topNum}{top X markers}

\item{plotHeatmap}{whether to plot heatmap or not (default yes)}

\item{group.order}{specify the cluster order (after pseudotime)}
}
\value{
a marker dataframe and the heatmap
}
\description{
marker identification by seurat
}
\examples{
markers <- marker_identification_by_seurat(seuset, topNum=10, plotHeatmap=T, group.order=NULL)
}
