% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_minPctChgTurnvoer.R
\name{minPctChgTurnover}
\alias{minPctChgTurnover}
\title{given a matrix of portfolio composition, minimize the turnover by only changing the position if target vs current is larger than certain percentage
does not work for negative weight
it does not guarantee wgt add up to 100percent}
\usage{
minPctChgTurnover(mat, thershold = 0.2, minPositionValue = 0.01,
  clearanceThershold = 0.005)
}
\arguments{
\item{mat}{matrix of turnover}

\item{thershold}{percentage of deviation between target and current}

\item{minPositionValue}{min position of an asset class, else will be zero}

\item{clearanceThershold}{if theoretical position fall below this thershold, then will clear the existing position}
}
\description{
given a matrix of portfolio composition, minimize the turnover by only changing the position if target vs current is larger than certain percentage
does not work for negative weight
it does not guarantee wgt add up to 100percent
}
