% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_minTurnover.R
\name{minTurnover}
\alias{minTurnover}
\title{given a matrix of portfolio composition, minimizee the portfolio turnover}
\usage{
minTurnover(mat, thershold = 0.005)
}
\description{
given a matrix of portfolio composition, minimizee the portfolio turnover
}
