% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_minTurnover.R
\name{minTurnovr}
\alias{minTurnovr}
\title{given a matrix of portfolio composition, minimizee the portfolio turnover}
\usage{
minTurnovr(mat, thershold = 0.05)
}
\description{
given a matrix of portfolio composition, minimizee the portfolio turnover
}

