% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{module_detection_along_pseudotime}
\alias{module_detection_along_pseudotime}
\title{pseudotime analysis based on clustering.}
\usage{
module_detection_along_pseudotime(exprM, annoDf, minNum = 100,
  maxNum = 200)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{startCluster}{starting Cluster}

\item{endCluster}{ending Cluster}
}
\value{
a new annoDf with pseudotime value for each cell
}
\description{
pseudotime analysis based on clustering.
}
\examples{
load("/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/Aug_30/analysis_4/Ctrl_YFP_ENCC.Rdata")
exprM <- seuset@scale.data; annoDf <- all_tsne
startCluster<-"Cluster1"; endCluster<-"Cluster6"
new_annoDf <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)
annoDf <- new_annoDf
}
