% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_na.locfCustom.R
\name{na.locfCustom}
\alias{na.locfCustom}
\title{custom funciton for na.locf, alter the behaviour of gap filling with paramter maxlocf
some parameter of na.locf is disabled due to imcompetibility with roll_sum}
\usage{
na.locfCustom(object, maxlocf = 0)
}
\arguments{
\item{object}{object that will be passed to na.locf/roll_sum.}

\item{maxlocf}{max number of days na.locf will be carried out. Default = 0 , do nothing}
}
\description{
custom funciton for na.locf, alter the behaviour of gap filling with paramter maxlocf
some parameter of na.locf is disabled due to imcompetibility with roll_sum
}
