% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{normalize_two_pseudotime}
\alias{normalize_two_pseudotime}
\title{normalize two pseudotime analysis, make sure the "common" are continuous, must at the beginning}
\usage{
normalize_two_pseudotime(annoDf, annoDf2, common)
}
\arguments{
\item{annoDf}{control annoDf}

\item{annoDf2}{to be aligned}

\item{common}{starting Cluster}
}
\value{
a new normalized annoDf2
}
\description{
normalize two pseudotime analysis, make sure the "common" are continuous, must at the beginning
}
\examples{
annoDf <- annoDf_glia; annoDf2 <- annoDf_neuron; common <- c("Cluster1", "Cluster4")
}
