% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolsets.R
\name{ora.go.kegg.clusterProfiler}
\alias{ora.go.kegg.clusterProfiler}
\title{GO KEGG ORA analysis}
\usage{
ora.go.kegg.clusterProfiler(geneList = markerList, organism = "hs")
}
\arguments{
\item{geneList}{a gene list}

\item{organism}{organism (hs and mm)}
}
\value{
a list with GO and KEGG annotation result (clusterProfiler format)
}
\description{
GO KEGG ORA analysis
}
\examples{
result <- ora.go.kegg.clusterProfiler(geneList = new_moduleList, organism="mm")

}
