% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCluster.R
\name{parallelCluster}
\alias{parallelCluster}
\title{parallelCluster}
\usage{
parallelCluster(n_cores = NULL, type = "PSOCK")
}
\arguments{
\item{n_cores}{number of cores}

\item{type}{type of cluster}
}
\value{
cluster object
}
\description{
parallelCluster
}
\examples{
cl <- parallelCluster()

}
