% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_msg.R
\name{pipe_msg}
\alias{pipe_msg}
\title{Mid message}
\usage{
pipe_msg(df, msg, type = NULL, time = T)
}
\arguments{
\item{df}{the df, to be passed with pipe}

\item{msg}{a message}

\item{type}{based con cli}
}
\value{
the original df so as to continue with the pipeline
}
\description{
Messages mid-pipeline
}
\examples{
# not yet

}
