% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.dotplot.human}
\alias{plot.dotplot.human}
\title{Draw dotplot for human SC3 object}
\usage{
\method{plot}{dotplot.human}(sce, genes.plot, use = "cellGroup2",
  xAngle = 60, cols.use = c("lightgrey", "blue"), col.min = -2.5,
  col.max = 2.5, dot.min = 0, dot.scale = 6, group.order = c(),
  scale.by = "radius", scale.min = NA, scale.max = NA, group.by,
  title = "", plot.legend = FALSE, do.return = FALSE,
  x.lab.rot = FALSE)
}
\arguments{
\item{sce}{SC3 object}

\item{genes.plot}{genes}

\item{use}{group used in SC3 object}

\item{xAngle}{angle of x title}

\item{cols.use}{colors used}

\item{col.min}{col.min}

\item{col.max}{col.max}

\item{dot.min}{dot.min}

\item{dot.scale}{whether to scale or not}

\item{group.order}{group.order}

\item{scale.by}{scale.by}

\item{scale.min}{scale.min}

\item{scale.max}{scale.max}

\item{group.by}{group.by}

\item{title}{title}

\item{plot.legend}{plot legend or not}

\item{do.return}{return object or not}

\item{x.lab.rot}{rotate x title or not}
}
\value{
a ggplot dotplot
}
\description{
Draw dotplot for human SC3 object
}
\examples{
options(repr.plot.width=4, repr.plot.height=6)
plot.dotplot.human(sce = sce_HSCR_pure, genes.plot = uniquegenes2, use="cellGroup", group.order = group.order, 
          scale.min=0, scale.max=100, title="", plot.legend = T, xAngle=90)

}
