% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.gsea.GO.KEGG.barplot.batch}
\alias{plot.gsea.GO.KEGG.barplot.batch}
\title{Draw GSEA GO KEGG barplot for a gene list}
\usage{
\method{plot}{gsea.GO.KEGG.barplot.batch}(anno_list = gsea_list$go_list,
  type)
}
\arguments{
\item{anno_list}{a gene list}

\item{type}{type of the input data}
}
\value{
a dataframe
}
\description{
Draw GSEA GO KEGG barplot for a gene list
}
\examples{
options(repr.plot.width=10, repr.plot.height=6)
go_list <- plot.gsea.GO.KEGG.barplot.batch(gsea_list$go_list, type = "GO")
kegg_list <- plot.gsea.GO.KEGG.barplot.batch(gsea_list$kegg_list, type = "KEGG")

}
