% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.violin}
\alias{plot.violin}
\title{draw violin plot}
\usage{
\method{plot}{violin}(exprData, cellAnno, genes, use, direction = "h",
  showGroup = T, groupOrder = NULL)
}
\arguments{
\item{exprData}{the expression matrix to be used}

\item{cellAnno}{the annotation of the cells}

\item{genes}{genes wanted to show}

\item{use}{the group selected in the cellAnno}

\item{direction}{direction of the violin plot}

\item{showGroup}{whether to show the group names}

\item{groupOrder}{the order of the group}
}
\value{
a ggplot violin object
}
\description{
draw violin plot
}
\examples{
options(repr.plot.width=4, repr.plot.height=9)
plot.violin(exprData=seuset@data, cellAnno=all_tsne_2, genes=key_genes, use="final", direction = "h")

all_tsne_3 <- subset(all_tsne_2, final \%in\% c('ctrl:c4','kif7:c4','ctrl:c6','kif7:c6'))
options(repr.plot.width=4, repr.plot.height=4)
plot.violin(exprData=seuset@data, cellAnno=all_tsne_3, genes=key_genes, use="final", direction = "h", 
            groupOrder = c('ctrl:c4','kif7:c4','ctrl:c6','kif7:c6'))

}
