% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEigengene.R
\name{plotEigengene}
\alias{plotEigengene}
\title{Plot Eigengene}
\usage{
plotEigengene(
  data,
  genes,
  yVector,
  xVector,
  xUnits = "Time",
  inverse = F,
  title = "",
  grid = F,
  colors = c("#009444", "#BE9230"),
  legend = T,
  multiline = F,
  legendTitle = "",
  plotBothProfiles = F
)
}
\arguments{
\item{data}{Samples in rows, genes in columns}

\item{genes}{list of genes to be represented}

\item{plotBothProfiles}{}
}
\description{
Description
}
