% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{given a matrix or xts object, ggplot the series}
\usage{
plotMatrix(mat, title = paste("plot time:", Sys.time()))
}
\description{
given a matrix or xts object, ggplot the series
}
