% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plotSingleTS.R
\name{plotSingleTS}
\alias{plotSingleTS}
\title{this function will be deprecated
wrapper for ggplot}
\usage{
plotSingleTS(ts, title)
}
\description{
this function will be deprecated
wrapper for ggplot
}

