% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting_genes_smooth_curve_across_pseudotime}
\alias{plotting_genes_smooth_curve_across_pseudotime}
\title{plotting smooth curve of different modules across pseudotime (in multiple samples)}
\usage{
plotting_genes_smooth_curve_across_pseudotime(exprM, annoDf, genes)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{moduleDf}{the marker genes (modules) dataframe}

\item{sampleCol}{sample column (for comparison)}
}
\value{
a ggplot object
}
\description{
plotting smooth curve of different modules across pseudotime (in multiple samples)
}
\examples{
HH_genes <- c("Shh","Dhh","Ptch1","Smo","Cdon","Gli1","Gli2","Gli3","Hhip","Spopl","Btrc","Cul1","Cul3","Sufu","Kif7")
HH_genes2 <- c( "Arrb2","Smurf1","Csnk1g3","Evc2","Csnk1g1","Ccnd2","Spop","Prkaca","Prkacb","Kif2a", "Gpr161","Csnk1g2","Csnk1a1","Fbxw11","Csnk1d","Csnk1e","Gsk3b","Boc","Evc","Bcl2","Arrb1","Gas1",  "Ccnd1","Smurf2")

genes <- HH_genes
}
