% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting_modules_smooth_curve_across_pseudotime}
\alias{plotting_modules_smooth_curve_across_pseudotime}
\title{plotting smooth curve of different modules across pseudotime (in multiple samples)}
\usage{
plotting_modules_smooth_curve_across_pseudotime(exprM, annoDf, moduleDf,
  sampleCol = NULL)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{moduleDf}{the marker genes (modules) dataframe}

\item{sampleCol}{sample column (for comparison)}
}
\value{
a ggplot object
}
\description{
plotting smooth curve of different modules across pseudotime (in multiple samples)
}
\examples{
load("/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/Aug_30/analysis_4/Ctrl_YFP_ENCC.Rdata")
exprM <- seuset@scale.data; annoDf <- all_tsne; moduleDf <- markers
startCluster<-"Cluster1"; endCluster<-"Cluster6"
result <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)
annoDf <- result
plotting_modules_smooth_curve_across_pseudotime()
}
