% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting_violin_plot_of_markers_across_clusters_by_seurat}
\alias{plotting_violin_plot_of_markers_across_clusters_by_seurat}
\title{plotting violin plot of markers across clusters}
\usage{
plotting_violin_plot_of_markers_across_clusters_by_seurat(seuset, markers,
  topNum = 3, cols.use = NULL)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{moduleDf}{the marker genes (modules) dataframe}

\item{sampleCol}{sample column (for comparison)}
}
\value{
a ggplot object
}
\description{
plotting violin plot of markers across clusters
}
\examples{
cols.use <- mycolors
plotting_violin_plot_of_markers_across_clusters_by_seurat(seuset, markers, topNum=3, cols.use=brewer.pal(8,"Set2"))
}
