% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plotTS.R
\name{plotTS}
\alias{plotTS}
\title{this function will be deprecated. 
wrapper for ggplot}
\usage{
plotTS(ts, title, additional = F, legend = "fx", ...)
}
\description{
this function will be deprecated. 
wrapper for ggplot
}

