% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolsets.R
\name{pre.pheatmap.anno}
\alias{pre.pheatmap.anno}
\title{prepare the annotation for the pheatmap}
\usage{
pre.pheatmap.anno(sce = sce, group.order = group.order,
  sample.order = sample.order, Cluster = "impute_cluster",
  Sample = "cellGroup")
}
\arguments{
\item{sce}{sce object}

\item{group.order}{group.order}

\item{sample.order}{sample.order}

\item{Cluster}{the feature take as Cluster}

\item{Sample}{the feature take as Sample}
}
\value{
anno list
}
\description{
prepare the annotation for the pheatmap
}
\examples{
options(repr.plot.width=8, repr.plot.height=7)
pheatmap(scale.data[sc3_marker$name, rownames(anno$col)], cluster_rows = F, cluster_cols = F, border_color = NA,
         color = colorRampPalette(c("#FF00FF","#000000","#FFFF00"))(100),
         show_colnames=F, show_rownames=F, annotation_col = anno$col, annotation_colors = anno$colors)

}
