% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toolsets.R
\name{pre.scale.data}
\alias{pre.scale.data}
\title{scale the matrix}
\usage{
pre.scale.data(raw.data = logcounts(sce_HSCR), max.value = 2)
}
\arguments{
\item{raw.data}{raw.data}

\item{max.value}{max.value after scale}
}
\value{
scale.data
}
\description{
scale the matrix
}
\examples{
scale.data <- pre.scale.data(logcounts(sce_HSCR))

}
