% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_matrix.R
\name{prepare_raw_count_matrix_from_10x_data}
\alias{prepare_raw_count_matrix_from_10x_data}
\title{prepare_raw_count_matrix_from_10x_data}
\usage{
prepare_raw_count_matrix_from_10x_data(cellranger_pipestance_path)
}
\arguments{
\item{cellranger_pipestance_path}{cellranger_pipestance_path}
}
\value{
a gbm object
}
\description{
prepare_raw_count_matrix_from_10x_data
}
