% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{pseudotime_based_on_clustering2}
\alias{pseudotime_based_on_clustering2}
\title{pseudotime analysis based on control clustering result.}
\usage{
pseudotime_based_on_clustering2(exprM_ctrl, exprM_all, annoDf,
  startCluster, endCluster, threads = 3)
}
\arguments{
\item{exprM_ctrl}{normalized expression matrix}

\item{exprM_all}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{startCluster}{starting Cluster}

\item{endCluster}{ending Cluster}
}
\value{
a new annoDf with pseudotime value for each cell
}
\description{
pseudotime analysis based on control clustering result.
}
\examples{
load("")
exprM_ctrl <- seuset@scale.data; exprM_all <- scale_data_by_seurat(merged_five_mouse_ENCC); annoDf <- new_annoDf
startCluster<-"Cluster1"; endCluster<-"Cluster6"
new_annoDf <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)
}
