% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdate.R
\name{rdate}
\alias{rdate}
\title{Prettier date}
\usage{
rdate()
}
\value{
Current date in a better format.
}
\description{
Sys.Date() but in a more familiar format and without unnecessary extras. dd-mm-yy
}
\examples{
'rdate()'
}
