% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_rdsToList.R
\name{rdsToList}
\alias{rdsToList}
\title{this function will be deprecated
given fieldname , retrieve the rds file form the corresponding directory}
\usage{
rdsToList(fieldname, filename = NULL)
}
\description{
this function will be deprecated
given fieldname , retrieve the rds file form the corresponding directory
}

