% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest2shp.R
\name{rest2shp}
\alias{rest2shp}
\title{REST service to shapefile}
\usage{
rest2shp(serviceURL, path)
}
\arguments{
\item{serviceURL}{URL of the REST service. Must be type FeatureServer. 'https://example.com/server/rest/services/someFolder/someWebApp/FeatureServer'}

\item{path}{Where to save the layers. 'D:/example/saveDir'}
}
\value{
Saves shapefiles to path
}
\description{
Saves an ArcGIS web layer as a shapefile.
}
\examples{
# rest2shp('https://example.com/server/rest/services/someFolder/someWebApp/FeatureServer', 'D:/example/saveDir')
}
