% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoolbox-package.R
\docType{package}
\name{rtoolbox-package}
\alias{rtoolbox}
\alias{rtoolbox-package}
\title{rtoolbox: Utilitaries R Functions}
\description{
A collection of miscellaneous R functions for everyday life.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/rtoolbox}
  \item Report bugs at \url{https://github.com/ahasverus/rtoolbox/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
