\name{ENCODER}
\alias{ENCODER}
\alias{list.files}
\title{
ENCODER: ENhanced COpy number Detection from Exome Reads
}
\description{
Generates DNA copy number profiles from whole exome sequence data using off-target reads
}
\usage{
ENCODER(bamFolder, destinationFolder, referenceFolder, whichControl, 
        ncpu, captureRegionsBedFile)
}

\arguments{
  \item{bamFolder}{the path to the folder containing all the .bam files that are to be included in the analysis. The path can be either absolute or relative.}
  \item{destinationFolder}{the path to the folder to which output should be written. The path can be either absolute or relative.}
  \item{referenceFolder}{the path to the folder with the helper files generated by \code{\link{preENCODER}}(). The helper files include the bin, mapability, GC-content, and blacklist files in .bed format}
  \item{whichControl}{the file that indicates which reference is to be used for every sample. WhichControl uses the order of the list of .bam files as presented by \code{\link{list.files}}(), and references are indicated by their indices. For instance, if the file list is Ref1.bam; Ref2.bam; Tumor1.bam; Tumor2.bam, and Ref1.bam (index 1) should be used as reference for Ref1.bam and Tumor1.bam, and Ref2.bam is the reference for Ref2.bam and Tumor2.bam, then c(1, 2, 1, 2) should be used.}
  \item{ncpu}{the number of CPUs that should be used for calculations.}
  \item{captureRegionsBedFile}{optional; the path to the capture design file, which should be in .bed format. If included, statistics on the overlap of peaks called by MACS and the capture regions will be provided.}
}
\details{
ENCODER (ENhanced COpy number Detection from Exome Reads) uses the off-target sequence reads from whole-exome sequencing or other targeted sequencing approaches to create copy number profiles. First, it removes non-random off-target reads, and it subsequently calculates the depth of coverage for the bins that are provided in the helper files. It then performs GC-content and mapability corrections, and removes blacklisted regions. \code{\link{plotCNA}}() generates a DNA copy number profile from the output of the \code{\link{ENCODER}}() function. The helper files can be created using \code{\link{preENCODER}}().

}
\value{
  \item{BamBaiMacsFiles}{a folder with the .bam, .bai and MACS files that are created during the \code{\link{ENCODER}}() run.}
  \item{Call_plots}{a folder that contains all copy number profile images.}
  \item{calls.Rdata}{an R object that contains all CGHcall call objects.}
  \item{input.Rdata}{an R object that contains information for \code{\link{plotCNA}}().}
  \item{log.txt}{log file of ENCODER.}
  \item{log2ratio_compensated_corrected.txt}{the file that contains the compenstaed corrected read counts after GC-content and mapability corrections, and after removal of data points in blacklisted regions. Counts are log2-transformed.}
  \item{qc}{a folder with quality control files. The .png files contain the plots and the loesses that are used for GC-content and mapability corrections. The fraction_of-bin .pdf files display the empirical cumulative distribution function for the fraction of bin (the  bin size after removal of peak regions expressed as a fraction of the original size.}
  \item{read_counts_compensated.txt}{the file that contains the raw and compensated read counts per bin.}
}
\references{
ENCODER: DNA copy number detection from off-target sequence data. Thomas Kuilman, Arno Velds, Kristel Kemper, Marco Ranzani, Lorenzo Bombardelli, Guotai Xu, Julian de Ruiter, Bauke Ylstra, Jos Jonkers, Sven Rottenberg, Lodewyk F. Wessels, David J. Adams, Daniel S. Peeper, Oscar Krijgsman. Submitted for publication. 
}
\author{
Thomas Kuilman (t.kuilman@nki.nl)\cr
Oscar Krijgsman (o.krijgsman@nki.nl)
}

\examples{
## Generate DNA copy number data from whole exome sequence data with annotation as generated by \code{\link{preENCODER}}.

\dontrun{ENCODER("./PATH_TO_BAMFOLDER/", "./PATH_TO_OUTPUTFOLDER/", "./PATH_TO_FOLDER/hg19_20kb/", rep(2, 3), 3)} ## No captureRegionsBedFile
\dontrun{ENCODER("./PATH_TO_BAMFOLDER/", "./PATH_TO_OUTPUTFOLDER/", "./PATH_TO_FOLDER/hg19_20kb/", rep(2, 3), 3, "./PATH_TO_CAPTUREREGIONSBEDFILE/captureRegionsBedFile.bed")} ## CaptureRegionsBedFile included in arguments
}
