% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_matrix.R
\name{scale_data_by_seurat}
\alias{scale_data_by_seurat}
\title{add_gender_info_to_annoDf}
\usage{
scale_data_by_seurat(rawCountM)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}
}
\value{
a ggplot object
}
\description{
add_gender_info_to_annoDf
}
\examples{

}
