% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpSize.R
\name{shpSize}
\alias{shpSize}
\title{Size of shapefile}
\usage{
shpSize(shapefile)
}
\arguments{
\item{shapefile}{Name of the shapefile.}
}
\value{
Size of shapefile bundle
}
\description{
Size of shapefile bundle, including all extensions that make up a shapefile. Useful for reporting.
}
\examples{
  shpSize('example.shp')
}
