% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent.R
\name{silent}
\alias{silent}
\title{Silent -- no messages or warnings}
\usage{
silent(expr)
}
\arguments{
\item{expr}{}
}
\value{
expr with no messages or warnings
}
\description{
Silent -- no messages or warnings
}
\examples{
# none
}
