% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_does_intersect.R
\name{st_does_intersect}
\alias{st_does_intersect}
\title{Does intersect}
\usage{
st_does_intersect(x, y)
}
\arguments{
\item{x}{A sf object}

\item{y}{A sf object}
}
\value{
A logical value
}
\description{
Checks if two sf objects intersect. Simplified return of `sf::st_intersects()`.
}
\examples{
# not yet
}
