% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{supervised_clustering}
\alias{supervised_clustering}
\title{supervised clustering based on distance (must have the same scale level for distance calculation)}
\usage{
supervised_clustering(exprM_to_be_c, exprM_ctrl, annoDf)
}
\arguments{
\item{exprM_to_be_c}{raw expression matrix from 10x (can't apply to smart-seq data)}

\item{exprM_ctrl}{known clustering annoDf}

\item{annoDf}{known clustering annoDf}

\item{colors.use}{colors used for clustering (should be more than the cluster num)}

\item{genes}{genes wanted to used in staining plot (FeaturePlot)}

\item{nCol}{column set in staining plot (FeaturePlot)}
}
\value{
a annoDf with clustering result
}
\description{
supervised clustering based on distance (must have the same scale level for distance calculation)
}
\examples{
exprM_to_be_c <- exprM_pos_neg
annoDf <- annoDf[colnames(exprM_ctrl),]
}
