% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-themes.R
\name{theme_bas}
\alias{theme_bas}
\title{Themes for ggplot2}
\usage{
theme_bas(
  base_size = 14,
  base_line_size = base_size/28,
  base_rect_size = base_size/28,
  grid = "XY"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{grid}{panel grid to be shown ("none" or a combination of "X", "Y", "x", and "y")}
}
\description{
Themes for ggplot2
}
