% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tag.R
\name{time_tag}
\alias{time_tag}
\title{Time tag}
\usage{
time_tag(full_date = F, full_time = F)
}
\arguments{
\item{full_date}{Whether to include year}

\item{full_time}{Whether to include seconds}
}
\value{
time tag in format ddMMyyhhmmss
}
\description{
Time tag
}
\examples{
# not yet
}
