% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unselect.R
\name{unselect}
\alias{unselect}
\title{Unselect}
\usage{
unselect(obj, x)
}
\arguments{
\item{obj}{expects df object}

\item{x}{vector of variables to unselect (string names)}
}
\value{
df without the variables
}
\description{
Unselect
}
\examples{
# not yet
}
