% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_updateByBDH.R
\name{updateByBDH}
\alias{updateByBDH}
\title{will be deprecate
function to download BDH data and store into rds}
\usage{
updateByBDH(field, CshAdjNormal = F, CshAdjAbnormal = F, CapChg = F,
  period = "DAILY", override_fields = NULL, override_values = NULL,
  altname = NULL)
}
\description{
will be deprecate
function to download BDH data and store into rds
}
