% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdEnsure.R
\name{wdEnsure}
\alias{wdEnsure}
\title{Ensure working directory}
\usage{
wdEnsure(wd)
}
\arguments{
\item{wd}{Working directory}
}
\value{
Message indicating the result of setting the working directory.
}
\description{
Make sure that the working directory is set correctly. It aims to eliminate the need to explicitly write an `if` statement and can replace `setwd()`.
}
\examples{
  wd <- 'C:/some_directory'
  wdEnsure(wd)
}
