% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x100of.R
\name{x100of}
\alias{x100of}
\title{Percent of}
\usage{
x100of(p, t, d = 2)
}
\arguments{
\item{p}{Desired percentage.}

\item{t}{Total.}

\item{d}{Digits to round to.}
}
\value{
x as a percentage of t
}
\description{
Some percentage of a total.
}
\examples{
# x100of(25, 200)

}
