% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip0.R
\name{zip0}
\alias{zip0}
\title{Zip but skip non-existant files}
\usage{
zip0(zipfile, files, overwrite = TRUE, mode = "mirror")
}
\arguments{
\item{zipfile}{}

\item{files}{}

\item{overwrite}{}

\item{mode}{}
}
\description{
Zip but skip non-existant files
}
\examples{
# ''
}
