% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipShp.R
\name{zipShp}
\alias{zipShp}
\title{Zip shapefile}
\usage{
zipShp(shapefile, shpDir, zipDir)
}
\arguments{
\item{shapefile}{Name of the shapefile. This will also be the name of the zipfile. 'example.shp'}

\item{shpDir}{Where to find the shapefile. 'D:/example/shpLocation'}

\item{zipDir}{Where to save the .zip file. 'D:/example/output'}
}
\value{
Saves zipped shapefile to shpDir.
}
\description{
Starting from a .shp file, finds all the files (.cpg, .dbf, .prj, .sbn, .sbx, .shp, .shp.xml, .shx) with the same name and zips them together to a .zip.
}
\examples{
# zipShp('someShp.shp', 'D:/example/shpLocation', 'D:/example/output')
# Would create D:/example/output/someShp.zip from:
# - someShp.cpg
# - someShp.dbf
# - someShp.prj
# - someShp.sbn
# - someShp.sbx
# - someShp.shp
# - someShp.shp.xml
# - someShp.shx
}
