% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_x100.R
\name{as_x100}
\alias{as_x100}
\title{As a percentage}
\usage{
as_x100(x, t, d = 2)
}
\arguments{
\item{x}{Value.}

\item{t}{Total.}

\item{d}{Digits to round to.}
}
\value{
x as a percentage of t
}
\description{
Takes a value and a total and returns the value as a percentage of that total.
}
\examples{
# as_x100(25, 200, 3)
}
