% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_combineFutSeries.R
\name{combineFutSeries}
\alias{combineFutSeries}
\title{given a list of futres chain series , concatenate them into one single time serie}
\usage{
combineFutSeries(futSeries)
}
\description{
given a list of futres chain series , concatenate them into one single time serie
}

