% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShp.R
\name{findShp}
\alias{findShp}
\title{Find shapefile}
\usage{
findShp(path)
}
\arguments{
\item{path}{Where to find the shapefiles. 'D:/example/location'}
}
\value{
Vector of shapefiles (only .shp).
}
\description{
Get a list (vector) of .shp files in a directory. Useful when looping over a directory containing all file types that make a shapefile. It allows to easily list and loop over only the .shp files, e.g. when using `sf::st_read()`.
}
\examples{
  findShp('C:/example/location')
}
