% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTif.R
\name{findTif}
\alias{findTif}
\title{Find TIFs}
\usage{
findTif(path, full = F)
}
\arguments{
\item{path}{Where to find the TIFs}
}
\value{
Vector of TIFs
}
\description{
Get a list (vector) of .tif files in a directory. Useful when looping over a directory that contains .tif files that come with .aux files.
}
\examples{
  findTif('C:/example/location')
}
