% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rid.R
\name{get_rid}
\alias{get_rid}
\title{Get rid of variable}
\usage{
get_rid(df, var.name)
}
\arguments{
\item{df}{data.frame object.}

\item{var.name}{Name (or names with c()) of variables to be dropped}
}
\value{
df without the specified variables.
}
\description{
Get rid of variables in less steps.
}
\examples{
# df <- df |> get_rid('variableName')
}
