% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpf.R
\name{gsea.go.kegg.clusterProfiler}
\alias{gsea.go.kegg.clusterProfiler}
\title{GSEA analysis by newest fgsea}
\usage{
gsea.go.kegg.clusterProfiler(geneList = DEGs_list_full,
  use.score = "cor", organism = "hs", pvalueCutoff = 0.05)
}
\arguments{
\item{geneList}{a gene list}

\item{use.score}{the score used for the GSEA analysis}

\item{organism}{organism (hs and mm)}

\item{pvalueCutoff}{pvalueCutoff}
}
\value{
a list with GO and KEGG GSEA annotation result (clusterProfiler format)
}
\description{
GSEA analysis by newest fgsea
}
\examples{
gsea_list <- gsea.go.kegg.clusterProfiler(geneList = pheno_DEGs, use.score = "cor", organism="mm")

}
