% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerFiles.R
\name{layerFiles}
\alias{layerFiles}
\title{File that make a shapefile}
\usage{
layerFiles(shapefile, full.names = T)
}
\arguments{
\item{shapefile}{Name of the shapefile in question.}
}
\value{
Vector
}
\description{
File that make a shapefile
}
\examples{
# layerFiles()
}
