% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_minPctChgTurnvoer.R
\name{minPctChgTurnover}
\alias{minPctChgTurnover}
\title{given a matrix of portfolio composition, minimize the turnover by only changing the position if target vs current is larger than certain percentage
does not work for negative weight
it does not guarantee wgt add up to 100%}
\usage{
minPctChgTurnover(mat, thershold = 0.2, minvalue = 0.005)
}
\arguments{
\item{mat}{matrix of turnover}

\item{thershold}{percentage of deviation between target and current}

\item{minvalue}{min size of change in any case}
}
\description{
given a matrix of portfolio composition, minimize the turnover by only changing the position if target vs current is larger than certain percentage
does not work for negative weight
it does not guarantee wgt add up to 100%
}

