% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{pass}
\alias{pass}
\title{Nothing}
\usage{
pass()
}
\description{
Literally does nothing. Mimics Python's 'pass' statement and just makes it a bit more familiar for someone that's used to Py.
}
\examples{
# if (1 > 2) {print('no')} else {pass()}
}
