% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting_genes_smooth_curve_across_pseudotime}
\alias{plotting_genes_smooth_curve_across_pseudotime}
\title{plotting smooth curve of different modules across pseudotime (in multiple samples)}
\usage{
plotting_genes_smooth_curve_across_pseudotime(exprM, annoDf, genes)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{moduleDf}{the marker genes (modules) dataframe}

\item{sampleCol}{sample column (for comparison)}
}
\value{
a ggplot object
}
\description{
plotting smooth curve of different modules across pseudotime (in multiple samples)
}
\examples{
HH_genes <- c("Smo", "Arrb2", "Smurf1", "Btrc", "Csnk1g3", "Evc2","Spopl", "Ptch1", "Csnk1g1", "Ccnd2", "Spop", "Prkaca", "Prkacb", "Kif2a", "Gpr161", "Sufu", "Gli2", "Cul1","Csnk1g2", "Csnk1a1", "Cul3","Fbxw11", "Csnk1d","Csnk1e","Gsk3b","Kif7","Boc","Cdon","Gli3","Evc","Bcl2","Arrb1","Hhip","Gas1", "Gli1", "Dhh","Shh","Ccnd1","Smurf2")
genes <- HH_genes
}
