% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting_modules_smooth_curve_across_pseudotime}
\alias{plotting_modules_smooth_curve_across_pseudotime}
\title{plotting smooth curve of different modules across pseudotime (in multiple samples)}
\usage{
plotting_modules_smooth_curve_across_pseudotime(exprM, annoDf, moduleList)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{moduleDf}{the marker genes (modules) dataframe}

\item{sampleCol}{sample column (for comparison)}
}
\value{
a ggplot object
}
\description{
plotting smooth curve of different modules across pseudotime (in multiple samples)
}
\examples{
load("/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/Aug_30/analysis_4/Ctrl_YFP_ENCC.Rdata")
exprM <- seuset@scale.data; annoDf <- all_tsne
startCluster<-"Cluster1"; endCluster<-"Cluster6"
new_annoDf <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)
annoDf <- new_annoDf
moduleDf <- markers
plotting_modules_smooth_curve_across_pseudotime(exprM, annoDf, moduleDf)

exprM <- exprM_all; annoDf <- new_annoDf; moduleDf <- markers; annoDf$cellGroup <- substr(rownames(annoDf), 1, 4)
1. seuset_3 <- SubsetData(object = seuset, ident.use = c("Cluster1", "Cluster2", "Cluster6"))
exprM <- seuset_3@scale.data
all_tsne <- TSNEPlot(object = seuset_3)$data[[1]]
rownames(all_tsne) <- colnames(seuset_3@data)
all_tsne$cluster <- as.character(seuset_3@ident[rownames(all_tsne)])
all_tsne$cellGroup <- substr(rownames(all_tsne), 1, 4)
annoDf <- all_tsne
startCluster <- "Cluster1"; endCluster <- "Cluster6"
annoDf <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)

2. exprM <- seuset_2@scale.data
all_tsne <- TSNEPlot(object = seuset_2)$data[[1]]
rownames(all_tsne) <- colnames(seuset_2@data)
all_tsne$cluster <- as.character(seuset_2@ident[rownames(all_tsne)])
all_tsne$cellGroup <- substr(rownames(all_tsne), 1, 4)
annoDf <- all_tsne
startCluster <- "Cluster1"; endCluster <- "Cluster3"
annoDf <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)

}
