% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{pseudotime_based_on_clustering}
\alias{pseudotime_based_on_clustering}
\title{pseudotime analysis based on clustering.}
\usage{
pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster,
  threads = 3)
}
\arguments{
\item{exprM}{normalized expression matrix}

\item{annoDf}{the annotation of the cells (columns) of the exprM}

\item{startCluster}{starting Cluster}

\item{endCluster}{ending Cluster}
}
\value{
a new annoDf with pseudotime value for each cell
}
\description{
pseudotime analysis based on clustering.
}
\examples{
load("/Users/surgery/Project/HOME/1-projects/1.scRNA-seq/3-10x/Aug_30/analysis_4/Ctrl_YFP_ENCC.Rdata")
exprM <- seuset@scale.data
all_tsne <- TSNEPlot(object = seuset)$data[[1]]
rownames(all_tsne) <- colnames(seuset@data)
all_tsne$cluster <- as.character(seuset@ident[rownames(all_tsne)])
all_tsne$cellGroup <- substr(rownames(all_tsne), 1, 4)
annoDf <- all_tsne
startCluster<-"Cluster1"; endCluster<-"Cluster6"
new_annoDf <- pseudotime_based_on_clustering(exprM, annoDf, startCluster, endCluster)
}
