% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdate.R
\name{rdate}
\alias{rdate}
\title{Prettier date}
\usage{
rdate()
}
\value{
Current date in a simpler dd-mm-yy format.
}
\description{
Sys.Date() but in a more familiar format and without unnecessary extras. Useful when reporting to the console or writing metadata.
}
\examples{
  rdate()
}
