% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtime.R
\name{rtime}
\alias{rtime}
\title{Prettier time}
\usage{
rtime(secs = T)
}
\arguments{
\item{secs}{Include seconds? T/F}
}
\value{
Current time in simpler HH:mm or HH:mm:ss format.
}
\description{
Sys.Time() but in a more familiar format and without unnecessary extras. Useful when reporting to the console; really helps with understanding execution flow.
}
\examples{
  rtime()
}
