% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x100of.R
\name{x100of}
\alias{x100of}
\title{Percentage}
\usage{
x100of(p, t, d = 2)
}
\arguments{
\item{p}{percentage value}

\item{t}{total}

\item{d}{digits to round}
}
\value{
x as a percentage of t
}
\description{
Percentage
}
\examples{
x100(25, 200)

}
